requirements.txt
attrs==21.2.0
Flask==2.0.1
Flask-SocketIO==5.1.0
python-dotenv==0.19.0
